refactor: type 관리 통일 (타입 전용 폴더로 분리)
