+++ apps/shop/src/features/product/types/index.ts
@@ -3,38 +3,29 @@ export interface ProductType {
     title: string;
     description: string;
     price: number;
-    pricePerUnit: string;
+    pricePerUnit?: string;
     images: string[];
     tags: string[];
-    badges: {
+    badges: Array<{
         text: string;
         bgColor: string;
         textColor?: string;
-    }[];
+    }>;
     inStock: boolean;
-    coffeeSize: string;
-    aromaFeatures: string[];
-    bodyLevel: number;
-    bitterLevel: number;
-    acidLevel: number;
-    roastLevel: number;
-    quantity: number;
-    limitDescription: string;
-    additionalDescription: string;
-    // 제품 상세 정보
+    coffeeSize?: string;
+    aromaFeatures?: string[];
+    bodyLevel?: number;
+    bitterLevel?: number;
+    acidLevel?: number;
+    roastLevel?: number;
+    quantity?: number;
+    limitDescription?: string;
+    additionalDescription?: string;
     productDetails?: ProductDetails;
 }
 
 export interface ProductDetails {
-    foodType: string;
-    origin: string;
-    expiryInfo: string;
-    volume: string;
-    ingredients: string;
-    returnPolicy: string;
-    storageMethod: string;
-    packaging: string;
-    notices: string[];
+    detailText: string;
 }
 
 export interface ReviewType {
@@ -53,4 +44,9 @@ export interface RecommendedProductType {
     price: number;
     image: string;
     inStock: boolean;
+}
+
+export interface ProductCategoryType {
+    id: number;
+    label: string;
 }
